requirements.txt
django>=4.2
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
pandas>=2.0.0
xhtml2pdf>=0.2.11
ofxparse>=0.21
python-dateutil>=2.8.2
